feat: add "coming soon" tag for current products
A user interface change to indicate upcoming features. Any product in the dashboard that has a status of 'coming-soon' now gets a "Coming soon" tag. This is implemented for better transparency with the users, allowing them to know what new features or products are expected to be available in the near future.