feat: add function to create sample documents
Added a new utility function in misc.js to handle the creation of sample model documents. This function 'useCreateSampleDocument' takes file data as parameter and makes a POST request to the '/product/sample' endpoint.

This change was made in order to improve the form submission handling in HelpTrainModel.jsx. Previously, the form submission was simulated with a delay function. Now it submits the data as FormData to the new function. Also, added error handling to toast an error message in case of failure.