feat: add search params to RequireAuth redirection
In the RequireAuth component, the useSearchParams hook from react-router-dom is now used when checking user's authentication status. Instead of always redirecting to the current pathname on unsuccessful auth verification, it first checks for a 'from' query param. This allows the component to remember the intended page after a successful login or email verification.